relocate client bits to messaging
