+++ crates/runtimelib/src/jupyter/mod.rs
@@ -1,4 +1,3 @@
-pub mod client;
 pub mod dirs;
 pub mod kernelspec;
